fix: downsampling with scale factor

+++ models/model_utils.py
@@ -103,8 +103,8 @@ class segmenthead(nn.Module):
         out = self.conv2(self.relu(self.bn2(x)))
 
         if self.scale_factor is not None:
-            height = x.shape[-2] * self.scale_factor
-            width = x.shape[-1] * self.scale_factor
+            height = int(x.shape[-2] * self.scale_factor)
+            width = int(x.shape[-1] * self.scale_factor)
             out = F.interpolate(out,
                         size=[height, width],
                         mode='bilinear', align_corners=algc)
